+++ README.md
@@ -1,1 +1,239 @@
-# get-review-todo
+<a name="readme-top"></a>
+
+<!--
+HOW TO USE:
+This is an example of how you may give instructions on setting up your project locally.
+
+Modify this file to match your project and remove sections that don't apply.
+
+REQUIRED SECTIONS:
+- Table of Contents
+- About the Project
+  - Built With
+  - Live Demo
+- Getting Started
+- Authors
+- Future Features
+- Contributing
+- Show your support
+- Acknowledgements
+- License
+
+OPTIONAL SECTIONS:
+- FAQ
+
+After you're finished please remove all the comments and instructions!
+-->
+
+# My Todo List Web Application
+
+<!-- - [x] Linter configs for **hint**, **stylelint** and **eslint**
+- [x] Initialized package.json and add linter dependencies
+- [x] `node_modules` folder in [.gitignore](/.gitignore)
+- [x] Include [Miroverse README.md](https://github.com/microverseinc/readme-template/blob/master/README.md)
+
+> Remove this block when you submit the project. -->
+
+---
+
+<!-- <div align="center"> -->
+  <!-- You are encouraged to replace this logo with your own! Otherwise you can also remove it. -->
+  <!-- <img src="murple_logo.png" alt="logo" width="140"  height="auto" /> -->
+  <!-- <h3><b>Microverse README Template</b></h3> -->
+<!-- </div> -->
+
+<!-- TABLE OF CONTENTS -->
+
+# 📗 Table of Contents
+
+- [📖 About the Project](#about-project)
+  - [🛠 Built With](#built-with)
+    - [Tech Stack](#tech-stack)
+    - [Key Features](#key-features)
+  - [🚀 Live Demo](#live-demo)
+- [💻 Getting Started](#getting-started)
+  - [Setup](#setup)
+  - [Prerequisites](#prerequisites)
+  - [Install](#install)
+  - [Usage](#usage)
+  - [Run tests](#run-tests)
+  - [Deployment](#triangular_flag_on_post-deployment)
+- [👥 Authors](#authors)
+- [🔭 Future Features](#future-features)
+- [🤝 Contributing](#contributing)
+- [⭐️ Show your support](#support)
+- [🙏 Acknowledgements](#acknowledgements)
+- [❓ FAQ (OPTIONAL)](#faq)
+- [📝 License](#license)
+
+<!-- PROJECT DESCRIPTION -->
+
+# 📖 Todo List Application <a name="about-project"></a>
+
+Todo List is an application where you write the tasks that needs to be done.
+
+## 🛠 Built With <a name="built-with"></a>
+
+<!-- ### Tech Stack <a name="tech-stack"></a> -->
+
+HTML, CSS, JavaScript
+
+<!-- Features -->
+
+### Key Features <a name="key-features"></a>
+
+- **[Add TAsks]**
+- **[Edit TAsks]**
+- **[Mark tasks as completed]**
+- **[Delete a task]**
+- **[Delete all tasks]**
+
+<p align="right">(<a href="#readme-top">back to top</a>)</p>
+
+### Screenshoots
+
+![App Screenshoot](/src/Screenshot_1.png)
+![App Screenshoot](/src/Screenshot_2.png)
+![App Screenshoot](/src/Screenshot_3.png)
+
+<!-- LIVE DEMO -->
+
+## 🚀 Live Demo <a name="live-demo"></a>
+
+- [Live Demo Link](https://auducodes-todo-app.netlify.app/)
+
+<p align="right">(<a href="#readme-top">back to top</a>)</p>
+
+<!-- GETTING STARTED -->
+
+## 💻 Getting Started <a name="getting-started"></a>
+
+Clone this repository to your machine and you are go to go:
+
+https://github.com/TamaraDimov/To-Do-List.git
+
+### Prerequisites
+
+Web Browser
+
+<!--
+Example command:
+
+```sh
+ gem install rails
+```
+ -->
+
+### Setup
+
+Download the repo.
+
+<!--
+Example commands:
+
+```sh
+  cd my-folder
+  git clone git@github.com:myaccount/my-project.git
+```
+--->
+
+### Install
+
+Install this project with:
+
+cd To-Do-List
+
+npm install
+
+<!--
+Example command:
+
+```sh
+  cd my-project
+  gem install
+```
+--->
+
+<!-- ### Usage -->
+
+<!--
+Example command:
+
+```sh
+  rails server
+```
+--->
+
+<!-- ### Run tests -->
+
+<!--
+Example command:
+
+```sh
+  bin/rails test test/models/article_test.rb
+```
+--->
+
+<!-- ### Deployment -->
+
+<!--
+Example:
+
+```sh
+
+```
+ -->
+
+<p align="right">(<a href="#readme-top">back to top</a>)</p>
+
+<!-- AUTHORS -->
+
+## 👥 Authors <a name="authors"></a>
+
+👤 **Kehinde Audu**
+
+- GitHub: [@githubhandle](https://github.com/audukehinde)
+- LinkedIn: [LinkedIn](https://www.linkedin.com/in/kehinde-audu-a44926175/)
+- Twitter: [Twitter](@auducodes)
+
+<!-- <p align="right">(<a href="#readme-top">back to top</a>)</p> -->
+
+<!-- FUTURE FEATURES -->
+
+<!-- ## 🔭 Future Features <a name="future-features"></a> -->
+
+<p align="right">(<a href="#readme-top">back to top</a>)</p>
+
+## 🤝 Contributing <a name="contributing"></a>
+
+Contributions, issues, and feature requests are welcome!
+
+Feel free to check the [issues page](../../issues/).
+
+<p align="right">(<a href="#readme-top">back to top</a>)</p>
+
+<!-- SUPPORT -->
+
+## ⭐️ Show your support <a name="support"></a>
+
+<!-- > Write a message to encourage readers to support your project -->
+
+If you like this project kindly show you support
+
+<p align="right">(<a href="#readme-top">back to top</a>)</p>
+
+<!-- ACKNOWLEDGEMENTS -->
+
+## 🙏 Acknowledgments <a name="acknowledgements"></a>
+
+<!-- I would like to thank -->
+
+<!-- <p align="right">(<a href="#readme-top">back to top</a>)</p> -->
+
+<!-- LICENSE -->
+
+## 📝 License <a name="license"></a>
+
+This project is [MIT](./LICENSE) licensed.
+
+<p align="right">(<a href="#readme-top">back to top</a>)</p>
